able to read value from matriarch

+++ src/matriarch/matr-book.h
@@ -14,9 +14,9 @@ struct _MatrBookClass
 {
     ScBookClass parent_class;
 
-    int (*read_control) (snd_seq_t *seq, snd_seq_addr_t addr, uint8_t read_ack, uint32_t control_id, uint8_t *val);
+    int (*read_control) (snd_seq_t *seq, snd_seq_addr_t addr, uint8_t read_ack, uint8_t control_id, uint16_t *val);
 };
 
-int matr_book_read_control (MatrBook *self, uint32_t control_id, uint8_t *val);
+int matr_book_read_control (MatrBook *self, uint8_t control_id, uint16_t *val);
 
 G_END_DECLS
